# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Strong Hire API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/job-targets/{job_target_id}/gap-analysis": {
      "post": {
        "tags": [
          "gap analysis"
        ],
        "summary": "Start Gap Analysis",
        "description": "Start the gap analysis for a job and a resume. Free; rate limited per user (GA-4).",
        "operationId": "start_gap_analysis_job_targets__job_target_id__gap_analysis_post",
        "parameters": [
          {
            "name": "job_target_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Target Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GapAnalysisStart"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GapAnalysisOut"
                }
              }
            }
          },
          "404": {
            "description": "The job target, the resume or the analysis does not exist."
          },
          "409": {
            "description": "The job posting or the resume is still being read."
          },
          "429": {
            "description": "Rate limited. detail.code is 'rate_limited'; see Retry-After."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "gap analysis"
        ],
        "summary": "Get Gap Analysis",
        "description": "The latest gap analysis for the job (GA-1 to GA-3). 404 when none was started.",
        "operationId": "get_gap_analysis_job_targets__job_target_id__gap_analysis_get",
        "parameters": [
          {
            "name": "job_target_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Target Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GapAnalysisOut"
                }
              }
            }
          },
          "404": {
            "description": "The job target, the resume or the analysis does not exist."
          },
          "409": {
            "description": "The job posting or the resume is still being read."
          },
          "429": {
            "description": "Rate limited. detail.code is 'rate_limited'; see Retry-After."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/gap-analyses/{gap_analysis_id}": {
      "get": {
        "tags": [
          "gap analysis"
        ],
        "summary": "Get Gap Analysis By Id",
        "description": "One gap analysis run of the signed-in user (R1 Reports page), also for a deleted job.",
        "operationId": "get_gap_analysis_by_id_gap_analyses__gap_analysis_id__get",
        "parameters": [
          {
            "name": "gap_analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Gap Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GapAnalysisOut"
                }
              }
            }
          },
          "404": {
            "description": "The job target, the resume or the analysis does not exist."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Competency": {
        "type": "string",
        "enum": [
          "ownership",
          "impact",
          "collaboration",
          "conflict_handling",
          "learning_from_failure",
          "communication",
          "role_fit",
          "depth_of_experience",
          "judgment",
          "motivation",
          "team_fit",
          "scope_at_level",
          "technical_depth",
          "accuracy",
          "reasoning",
          "trade_offs",
          "clarity_of_explanation",
          "problem_framing",
          "structure",
          "user_and_business_sense",
          "estimation_logic",
          "prioritization",
          "recommendation"
        ],
        "title": "Competency",
        "description": "Scored competencies from the spec, grouped by interview type in COMPETENCIES_BY_TYPE."
      },
      "CompetencyMatch": {
        "properties": {
          "competency": {
            "$ref": "#/components/schemas/Competency"
          },
          "score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "competency",
          "score"
        ],
        "title": "CompetencyMatch"
      },
...
      "Difficulty": {
        "type": "string",
        "enum": [
          "friendly",
          "realistic",
          "tough"
        ],
        "title": "Difficulty",
        "description": "IV-4."
      },
...
      "Gap": {
        "properties": {
          "summary": {
            "type": "string",
            "minLength": 1,
            "title": "Summary"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "related_requirement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Related Requirement"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "summary",
          "severity"
        ],
        "title": "Gap"
      },
      "GapAnalysis": {
        "properties": {
          "match_score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Match Score",
            "description": "GA-1."
          },
          "requirement_breakdown": {
            "items": {
              "$ref": "#/components/schemas/RequirementMatch"
            },
            "type": "array",
            "title": "Requirement Breakdown"
          },
          "competency_breakdown": {
            "items": {
              "$ref": "#/components/schemas/CompetencyMatch"
            },
            "type": "array",
            "title": "Competency Breakdown"
          },
          "strengths": {
            "items": {
              "$ref": "#/components/schemas/Strength"
            },
            "type": "array",
            "title": "Strengths"
          },
          "gaps": {
            "items": {
              "$ref": "#/components/schemas/Gap"
            },
            "type": "array",
            "title": "Gaps"
          },
          "probe_areas": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Probe Areas"
          },
          "session_plan": {
            "items": {
              "$ref": "#/components/schemas/PlannedSession"
            },
            "type": "array",
            "minItems": 1,
            "title": "Session Plan"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "match_score",
          "session_plan"
        ],
        "title": "GapAnalysis"
      },
      "GapAnalysisOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "job_target_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Target Id"
          },
          "resume_id": {
            "type": "string",
            "format": "uuid",
            "title": "Resume Id"
          },
          "resume_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume Name",
            "description": "R1: the CV's library name. None when deleted."
          },
          "resume_deleted": {
            "type": "boolean",
            "title": "Resume Deleted",
            "description": "R1: the CV used for this analysis was deleted."
          },
          "job_deleted": {
            "type": "boolean",
            "title": "Job Deleted",
            "description": "R1: the job description was deleted."
          },
          "status": {
            "$ref": "#/components/schemas/GapStatus"
          },
          "analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GapAnalysis"
              },
              {
                "type": "null"
              }
            ],
            "description": "Set when status is ready."
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "A plain reason when status is failed."
          },
          "generic_mode": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generic Mode",
            "description": "True when no company profile was used. None until ready."
          },
          "profile_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Version"
          },
          "stale": {
            "type": "boolean",
            "title": "Stale",
            "description": "The job, the resume or the company profile changed after this analysis. Start it again to update it. Always false when the job or the resume was deleted."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "job_target_id",
          "resume_id",
          "resume_name",
          "resume_deleted",
          "job_deleted",
          "status",
          "analysis",
          "error",
          "generic_mode",
          "profile_version",
          "stale",
          "created_at",
          "updated_at"
        ],
        "title": "GapAnalysisOut",
        "description": "The latest gap analysis of a job target. Poll while status is \"running\"."
      },
      "GapAnalysisStart": {
        "properties": {
          "resume_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume Id",
            "description": "The resume to compare with the job. Leave it out to run again with the resume of the latest analysis."
          },
          "reuse_ready": {
            "type": "boolean",
            "title": "Reuse Ready",
            "description": "R1: return the latest ready analysis instead of a new run when the job, the resume and the company profile did not change since it was made. The app sets it when the user picks a saved job and a saved CV again.",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "GapAnalysisStart"
      },
...
      "GapStatus": {
        "type": "string",
        "enum": [
          "running",
          "ready",
          "failed"
        ],
        "title": "GapStatus",
        "description": "State of one gap analysis run (GA-1 to GA-4)."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InterviewType": {
        "type": "string",
        "enum": [
          "behavioral",
          "hiring_manager",
          "technical_qa",
          "case"
        ],
        "title": "InterviewType",
        "description": "The four v1 interview types (IV-2)."
      },
...
      "PlannedSession": {
        "properties": {
          "priority": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Priority",
            "description": "1 is the first session to run."
          },
          "interview_type": {
            "$ref": "#/components/schemas/InterviewType"
          },
          "difficulty": {
            "$ref": "#/components/schemas/Difficulty",
            "default": "realistic"
          },
          "focus_topics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Focus Topics"
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "priority",
          "interview_type",
          "focus_topics",
          "reason"
        ],
        "title": "PlannedSession",
        "description": "One recommended practice session (GA-3)."
      },
...
      "RequirementKind": {
        "type": "string",
        "enum": [
          "must_have",
          "nice_to_have"
        ],
        "title": "RequirementKind"
      },
      "RequirementMatch": {
        "properties": {
          "requirement": {
            "type": "string",
            "minLength": 1,
            "title": "Requirement"
          },
          "kind": {
            "$ref": "#/components/schemas/RequirementKind"
          },
          "score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score"
          },
          "evidence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence",
            "description": "Resume text that supports the score."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requirement",
          "kind",
          "score"
        ],
        "title": "RequirementMatch"
      },
...
      "Severity": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high"
        ],
        "title": "Severity"
      },
...
      "Strength": {
        "properties": {
          "summary": {
            "type": "string",
            "minLength": 1,
            "title": "Summary"
          },
          "evidence": {
            "type": "string",
            "minLength": 1,
            "title": "Evidence",
            "description": "Resume evidence (GA-2)."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "summary",
          "evidence"
        ],
        "title": "Strength"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}